+++ cmd/notation/internal/display/metadata/interface.go
@@ -11,6 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package metadata provides interfaces for handlers that render metadata
+// information for each command.
 package metadata
 
 import (
